+++ __init__.py
@@ -20,6 +20,9 @@ try:
     from .k_easy_utility import NODE_CLASS_MAPPINGS as utility_mappings
     from .k_easy_utility import NODE_DISPLAY_NAME_MAPPINGS as utility_display
 
+    from .k_audio_timeline import NODE_CLASS_MAPPINGS as audio_timeline_mappings
+    from .k_audio_timeline import NODE_DISPLAY_NAME_MAPPINGS as audio_timeline_display
+
     from .forks.radiance_koolook import NODE_CLASS_MAPPINGS as radiance_koolook_mappings
     from .forks.radiance_koolook import NODE_DISPLAY_NAME_MAPPINGS as radiance_koolook_display
 
@@ -85,6 +88,7 @@ else:
         **cam_loader_mappings,
         **pattern_mappings,
         **utility_mappings,
+        **audio_timeline_mappings,
         **video_load_mappings,
         **video_combine_mappings,
         **radiance_koolook_mappings,
@@ -99,6 +103,7 @@ else:
         **cam_loader_display,
         **pattern_display,
         **utility_display,
+        **audio_timeline_display,
         **video_load_display,
         **video_combine_display,
         **radiance_koolook_display,